First steps in join logic
This commit introduces small steps towards a proper
join logic, separating client and server properly.

Ie. the client connect to the local instance sending
it an anchor node. Then the local server picks it up,
builds a capnp client to contact the remote server
which will accept our membership, and sends a request
to the anchor node with informations and a capnp client
handle.

We need to ensure that Topology and Server are connected
via a channel, and that the join is made through the
Server, which is the only component able to build a
capnp Client. Similarly, once we receive the ClusterSync
answer, Server needs to send the information to other
components, such as the members, workloads, etc. via
channels.

This is the only way to keep things separate and clean.
Capnproto makes this really difficult...