store: Implement delta state diff streaming for peers
This commit adds a very rough base for streaming delta
state over the Crdt/MST store for Peers. After joining
the cluster, a node will get the Sync interface from
the remote anchor and then pull the list of peers.

This should be obviously generalized to all types of
objects and a background task should be syncing the
full state for all domains (peers/containers/networks/etc.).

For now we keep it constrained as an example. This code
needs refactoring and simplifying.