store: Solve root hash discrepancy due to MVReg duplicated entry
This commit fixes the root hash discrepancy after joining
a Node to a cluster. The anchor would send its own PeerValue
along with the joiner's own PeerValue, which would be added
to the MVReg without being deduplicated. This results in a
discrepancy on the value of the MVReg for the joiner node,
itself resulting in a diverging root_hash. We solve this
by deduplicating and only allowing a single value on the
MVReg.

While using a MVReg makes sense in case of concurrent entries
being spread around via gossip, it also makes it more complex
to solve anti-entropy at scale, since we could easily end up
with diverging root hashes for the Merke Search Tree and send
useless diffs around, flooding the network. Maybe we should
consider using a LWW register instead but we lose the capacity
to accept concurrent updates.

This needs to be figured out.