auth: Move token auth from Noise handshake to join capnp rpc logic
This commit moves the token authentication flow from
the Noise protocol layer to the capnproto rpc logic
layer. This is more in line with how capnproto handles
security via capabilities and simplifies the connection
logic. We pass on the auth token via capnproto on join/
registerNode and if the token doesn't match, we do not
return the Sync capability. If the token matches, we
return the Sync cap as part of the response and the
client can proceed to fetch data.