server: Refactor node registration logic
This commit does a large refactor of the node registration
logic that used to happen within the Topology component.
The main problem with the old capnproto interface structure
was that a node could bypass the token check and simply get
any interface (Topology/Gossip/Scheduling) it liked since
the join method was accessible only via Topology which had
to remain ungated. We need to ensure that there is a single
entrypoint for node registration and getting a ClusterSession
for access to all other capabilities. This is more in line
with capnproto usage.

To do this, we move the registerNode() method to Server and
return a ClusterSession on completion. We do initial work
to store local tickets to be able to resume the connection
on resumeSession(). Still needs more improvements in the
way one node will access all of the other peers with
cluster wide issued credentials. As of now, a node would
have to go through registerNode on each of the other peers
it gets via gossip/sync in order to be able to contact them.

Future work needs to figure that part out.