server: Restore connection to peers on restart and refresh capnp handles
This commit stores a local ticket on registerNode and implements
mutual trust between peers. It restores the connection to other
peers on restart, such that the capnp handles are fresh and
ready to use for gossiping and anti-entropy/topology operations.

There are improvements to make on the credentials path which is
used for Node A to connect to Node C after registering with Node B.
Since Node A doesn't have any ticket for Node C, but could use
credentials minted from Node B and that Node C could validate
with Node B's credentials after Node C knows about Node B.