server: Fix race with issuing reciprocal ticket on node join
This commit fixes an issue with the join process, where the
anchor node asks for a reciprocal ticket to the joining peer.
Since the anchor node information doesn't exist until after
the first sync completes, there could be a race happening
when the anchor asks for a reciprocal ticket and the joining
node rejects it since it doesn't have its peer id inside its
store.

We implement a workaround by inserting a dummy entry for the
peer with its peer id to be able to issue the ticket. This
entry is overridden by the following sync, which will fill
the rest of the information successfully.