tests: introduce root hash assertion to register tests
This commit introduces an assertion in register_node_*
tests to check for merkle search tree root-hashes being
equal. It also fixes a bug on the register tests where
the advertised address would be incompatible with local
tests, assigning the node's IP instead of overriding
for in process transport or loopback interface (local).

This resulted in nodes not being able to communicate with
each other in tests. The test was succeeding since we were
just checking on the length of the peers list and the
initial communication would succeed, but the sync would
then fail.