server: Refactor registerNode and getWithCredential to return NodeInfo
This commit refactors the registerNode and getWithCredential
methods to return the full NodeInfo of the anchor/remote node
instead of just the nodeId. This is helpful when joining since
the anchor needs to be in the joiner's Peers list and storage
in order to get assigned a ticket. Otherwise, the anchor needs
to wait for the next MST anti-entropy sync tick to be able to
get the ticket and contact spread informations to the joiner.