scheduler: First work around slot scheduling algorithm
This commit contains an initial version of the slot scheduling
algorithm with a basic scheduler. It only works locally for now.
Further work is required to reserve slots across nodes and
implement scheduling strategies. As it stands it is a very
naive approach and requires a lot of clean up.